1.增加歌单详情页（无歌曲列表） 2.测试共享元素跳转

+++ composeApp/src/commonMain/kotlin/com/mio/bean/OriginSongSimpleData.kt
@@ -13,5 +13,5 @@ data class OriginSongSimpleData(
     @SerialName("name")
     val name: String = "",
     @SerialName("songId")
-    val songId: Int = 0
+    val songId: Long = 0
 )
